fix: preserve token login flow and split email login API
